feat: rename files to fit pythonic conventions (modules should always be lowercase)
