+++ src/main.py
@@ -1,4 +1,4 @@
-from Window import Window
+from window import Window
 
 window = Window()
 window.show_menu_screen()
